Rename tiny_components.hpp to tarjan_scc.hpp
Fixes #1561
